% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sew.R
\name{SEW}
\alias{SEW}
\title{SEW}
\usage{
SEW(chr, posfile, outputdir, tempdir = NA, bamlist = "",
  cramlist = "", reference = "", genfile = "", phasefile = "",
  outputHaplotypeProbabilities = FALSE, regenerateInput = TRUE,
  originalRegionName = NA, keepInterimFiles = FALSE,
  keepTempDir = FALSE, generateInputOnly = FALSE,
  useSoftClippedBases = FALSE, vcf_output_name = NULL,
  initial_min_hapProb = 0.4, initial_max_hapProb = 0.6,
  regenerateInputWithDefaultValues = FALSE, unwindIterations = c(40,
  70, 100, 120, 140, 160, 180, 200), sample_unwindIterations = c(5, 5, 1,
  1, 1, 1, 1, 1), usePhaseSet_unwindIterations = c(rep(FALSE, 8)),
  save_sampleReadsInfo = TRUE, outputInputInVCFFormat = FALSE,
  downsampleToCov = 200, downsampleFraction = 1, chrStart = NA,
  chrEnd = NA, regionStart = NA, regionEnd = NA, buffer = NA,
  iSizeUpperLimit = as.integer(600), bqFilter = as.integer(-1),
  niterations = 300, plotHapSumDuringIterations = FALSE,
  very_verbose = FALSE, keepSampleReadsInRAM = FALSE)
}
\arguments{
\item{chr}{What chromosome to run. Should match BAM headers}

\item{posfile}{Where to find file with positions to run. File is tab seperated with no header, one row per SNP, with col 1 = chromosome, col 2 = physical position (sorted from smallest to largest), col 3 = reference base, col 4 = alternate base. Bases are capitalized. Example first row: 1<tab>1000<tab>A<tab>G<tab>}

\item{outputdir}{What output directory to use}

\item{tempdir}{What directory to use as temporary directory. If possible, use ramdisk, like /dev/shm/}

\item{bamlist}{Path to file with bam file locations. File is one row per entry, path to bam files. Bam index files should exist in same directory as for each bam, suffixed either .bam.bai or .bai}

\item{cramlist}{Path to file with cram file locations. File is one row per entry, path to cram files. cram files are converted to bam files on the fly for parsing into SEW}

\item{reference}{Path to reference fasta used for making cram files. Only required if cramlist is defined}

\item{genfile}{Path to gen file with high coverage results. Empty for no genfile. File has a header row with a name for each sample, matching what is found in the bam file. Each subject is then a tab seperated column, with 0 = hom ref, 1 = het, 2 = hom alt and NA indicating missing genotype, with rows corresponding to rows of the posfile. Note therefore this file has one more row than posfile which has no header}

\item{phasefile}{Path to phase file with truth phasing results. Empty for no phasefile. File has a header row with a name for each sample, matching what is found in the bam file. Each subject is then a tab seperated column, with 0 = ref and 1 = alt, separated by a vertical bar |, e.g. 0|0 or 0|1. Note therefore this file has one more row than posfile which has no header}

\item{outputHaplotypeProbabilities}{Whether to output haplotype probabilities in files}

\item{regenerateInput}{Whether to regenerate input files}

\item{originalRegionName}{If regenerateInput is FALSE (i.e. using existing data), this is the name of the original region name (chr.regionStart.regionEnd). This is necessary to load past variables}

\item{keepInterimFiles}{Whether to keep interim parameter estimates}

\item{keepTempDir}{Whether to keep files in temporary directory}

\item{generateInputOnly}{Whether to just generate input data then quit}

\item{useSoftClippedBases}{Whether to use (TRUE) or not use (FALSE) bases in soft clipped portions of reads}

\item{vcf_output_name}{Override the default VCF output name with this given file name. Please note that this does not change the names of inputs or outputs (e.g. RData, plots), so if outputdir is unchanged and if multiple SEW runs are processing on the same region then they may over-write each others inputs and outputs}

\item{initial_min_hapProb}{Initial lower bound for probability read comes from haplotype. Double bounded between 0 and 1}

\item{initial_max_hapProb}{Initial upper bound for probability read comes from haplotype. Double bounded between 0 and 1}

\item{regenerateInputWithDefaultValues}{If regenerateInput is FALSE and the original input data was made using regionStart, regionEnd and buffer as default values, set this equal to TRUE}

\item{unwindIterations}{What iterations to search for flips in estimated haplotypes that increase the likelihood}

\item{sample_unwindIterations}{When performing an unwind iterations, how often to check a SNP for an improvement following unwinding. 1 means check all SNPs}

\item{save_sampleReadsInfo}{Experimental. Boolean TRUE/FALSE about whether to save additional information about the reads that were extracted}

\item{outputInputInVCFFormat}{Whether to output the input in vcf format}

\item{downsampleToCov}{What coverage to downsample individual sites to. This ensures no floating point errors at sites with really high coverage}

\item{downsampleFraction}{Downsample BAMs by choosing a fraction of reads to retain. Must be value 0<downsampleFraction<1}

\item{chrStart}{When loading from BAM, some start position, before SNPs occur. Default NA will infer this from either regionStart, regionEnd and buffer, or posfile}

\item{chrEnd}{When loading from BAM, some end position, after SNPs occur. Default NA will infer this from either regionStart, regionEnd and buffer, or posfile}

\item{regionStart}{When running imputation, where to start from. The 1-based position x is kept if regionStart <= x <= regionEnd}

\item{regionEnd}{When running imputation, where to stop}

\item{buffer}{Buffer of region to perform imputation over. So imputation is run form regionStart-buffer to regionEnd+buffer, and reported for regionStart to regionEnd, including the bases of regionStart and regionEnd}

\item{iSizeUpperLimit}{Do not use reads with an insert size of more than this value}

\item{bqFilter}{Minimum BQ for a SNP in a read. Also, the algorithm uses bq<=mq, so if mapping quality is less than this, the read isnt used}

\item{niterations}{Number of EM iterations.}

\item{plotHapSumDuringIterations}{Boolean TRUE/FALSE about whether to make a plot that shows the relative number of individuals using each ancestral haplotype in each iteration}

\item{very_verbose}{Boolean TRUE/FALSE about whether to print additional values which may be helpeful for understanding how the model operates}

\item{keepSampleReadsInRAM}{Whether to (generally) keep sampleReads in RAM or store them in the temporary directory. SEW will be faster if this is FALSE at the expense of RAM}
}
\value{
Results in properly formatted version
}
\description{
SEW
}
\author{
Robert Davies
}
